feat: update imports and move setFocus to utils
