+++ packages/date-picker/docs/Calendar.stories.tsx
@@ -4,7 +4,7 @@ import { usePopper } from "react-popper"
 import { enAU } from "date-fns/locale"
 import { CATEGORIES, SUB_CATEGORIES } from "../../../storybook/constants"
 import { DayOfWeek } from "../src/enums"
-import { Calendar, CalendarProps } from "../src/_primitives/Calendar/Calendar"
+import { Calendar, CalendarProps } from "../src/_primitives/Calendar"
 import { StoryWrapper } from "../../../storybook/components/StoryWrapper"
 
 const SUPPORTED_LOCALES = ["en-US", "en-AU"]
