+++ packages/date-picker/index.ts
@@ -1,3 +1,4 @@
 export * from "./src/DatePicker"
 export * from "./src/DateRangePicker"
 export * from "./src/enums"
+export * from "./src/types"
