+++ packages/date-picker/src/utils/setFocusInCalendar.ts
@@ -1,6 +1,6 @@
-import calendarStyles from "../Calendar.scss"
-import { CalendarElement } from "../Calendar"
-import { isInvalidDate } from "../../../utils/isInvalidDate"
+import calendarStyles from "../_primitives/Calendar/Calendar.scss"
+import { CalendarElement } from "../_primitives/Calendar"
+import { isInvalidDate } from "./isInvalidDate"
 
 const isHTMLElement = (element: Element | undefined): element is HTMLElement =>
   element instanceof HTMLElement
